added authentication for login and reset password

+++ controllers/bootcamps.js
@@ -1,11 +1,9 @@
 const Bootcamp = require("../models/Bootcamp");
 const ErrorResponse = require("../utils/errorResponse");
 const asyncHandler = require("../middleware/async");
-const path = require('path')
+const path = require("path");
 
 exports.getBootcamps = asyncHandler(async (req, res, next) => {
-
-
   res.status(200).json(res.advancedResults);
 });
 
@@ -17,10 +15,28 @@ exports.getBootcampById = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp with id ${req.params.id} not found`, 404)
     );
   }
+
   res.status(200).json({ success: true, data: bootcamp });
 });
 
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+  //Add user to req.body
+  req.body.user = req.user.id;
+
+  // check for published bootcamp
+
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+  // If the user is not an admin, they can only add one bootcammp
+  if (publishedBootcamp && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `The user with Id ${req.user.id} has already published a bootcamp`,
+        400
+      )
+    );
+  }
+
   const bootcamp = await Bootcamp.create(req.body);
   res.status(201).json({
     success: true,
@@ -29,16 +45,25 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 });
 
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  const bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp with id ${req.params.id} not found`, 404)
     );
   }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `user with id ${req.params.id} not authorized to update`,
+        404
+      )
+    );
+  }
+
+  bootcamp = await Bootcamp.findOneAndUpdate(req.params.id, req.body);
   res.status(200).json({ success: true, data: bootcamp });
 });
 
@@ -49,9 +74,19 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp with id ${req.params.id} not found`, 404)
     );
   }
+
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `user with id ${req.params.id} not authorized to delete`,
+        404
+      )
+    );
+  }
   // Call remove() on the bootcamp instance
   await bootcamp.deleteOne();
-  res.status(200).json({ success: true, data: {} });
+  res.status(200).json({ success: true, msg: "bootcamp deleted successfully" });
 });
 
 // photo uploads for bootcamp
@@ -65,6 +100,15 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `user with id ${req.params.id} not authorized to delete`,
+        404
+      )
+    );
+  }
 
   if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 404));
@@ -75,7 +119,7 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     return next(new ErrorResponse("Please upload an image file", 400));
   }
 
-  //Check filesize
+  //Check file size
   if (file.size > process.env.MAX_FILE_UPLOAD) {
     return next(
       new ErrorResponse(
@@ -90,8 +134,8 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   console.log(file.name);
 
   //upload file
-  const url = `${req.get('host')}/uploads/${file.name}`
-  console.log(url)
+  const url = `${req.get("host")}/uploads/${file.name}`;
+  console.log(url);
   file.mv(`${process.env.FILE_UPLOAD_PATH}/${file.name}`, async (err) => {
     if (err) {
       console.log(err);
@@ -99,10 +143,10 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     }
     await Bootcamp.findByIdAndUpdate(req.params.id, { photo: file.name });
 
-    res.status(200).json ({
-      sucess:true,
+    res.status(200).json({
+      sucess: true,
       data: file.name,
-      url
-    })
+      url,
+    });
   });
 });
